kml2sct2.jar
Included a tool (written by me) to convert KML polygons into the sector2 format. Does not currently work anything that is not a polygon. I will likely update the program to work with both shape types as well as a "select file" option rather than having to type in the file name for ease of use.